Alter cleanup behavior, stage blocks instead of uploading
If we fail the transfer, outright delete the blob.

If we cancel the transfer, only delete the blob if there's not a committed blob present.

Make a full-file single chunk upload stage a block instead of using upload-- Doing this means that if the transfer gets cancelled in a single-chunk transfer, the deletion code will still trigger correctly.

This also has the side effect of making the code paths more consistent.